fix: find backwards broken
If the cursor is directly at the first letter of the word, that word's
instance should be ignored. Otherwise, we can never traverse backwards
in a search, because the textbox does not manually subtract one from
the cursor position (and it should never need to, since there can be
troublesome edge cases that should be handled by the text buffer, rather
than by the textbox).